app/requirements.txt
flask 
Flask-SQLAlchemy 
celery 
redis 
requests